requirements.txt
# Discord
discord.py>=2.3.2

# API Clients
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-generativeai>=0.7.2  # Gemini 2.5 Pro 지원 버전
pytrends>=4.9.2

# Database (PostgreSQL for backup)
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.3
aiohttp>=3.8.5
scikit-learn>=1.3.0  # 예측 엔진용

# Memory Caching
cachetools>=5.3.2
diskcache>=5.6.3

# Environment
python-dotenv>=1.0.0

# Utilities
dataclasses-json>=0.6.1
tenacity>=8.2.3
aiofiles>=23.2.1

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.1